backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.0
openai>=1.30.0
transformers>=4.40.0
torch>=2.2.0
pymongo>=4.7.0
motor>=3.4.0
httpx>=0.27.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
